requirements.txt
Django==5.1.7
djangorestframework==3.15.2
requests==2.32.5
asgiref==3.8.1
sqlparse==0.5.3
tzdata==2025.1
